% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{A4TempStatistic}
\alias{A4TempStatistic}
\title{generate the randomization statistic based on A4}
\usage{
A4TempStatistic(theData, S)
}
\arguments{
\item{theData}{the data}
}
\value{
ranStat a function of random group transformation
}
\description{
generate the randomization statistic based on A4
}
