% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{chenTempStatistic}
\alias{chenTempStatistic}
\title{generate the randomization statistic based on Chen's test statistic}
\usage{
chenTempStatistic(theData)
}
\arguments{
\item{theData}{the data}
}
\value{
ranStat a function of random group transformation
}
\description{
generate the randomization statistic based on Chen's test statistic
}
