% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{fanFactorModelGen}
\alias{fanFactorModelGen}
\title{generate data by factor model as Fan (2007)}
\usage{
fanFactorModelGen(n, p, theParam, mu)
}
\arguments{
\item{n}{n}

\item{p}{p}
}
\description{
generate data by factor model as Fan (2007)
}
