% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{genSpikedModel}
\alias{genSpikedModel}
\title{generate normal factor model}
\usage{
genSpikedModel(n, p, mu, modelParam)
}
\arguments{
\item{modelParam}{the parameters of spiked model, including theSigmaSq, V, D, r}
}
\description{
generate normal factor model
}
