% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{kSamplePermutation}
\alias{kSamplePermutation}
\title{generate k sample permutation}
\usage{
kSamplePermutation(k, Ns, ran = sample.int(sum(Ns)))
}
\value{
a list of k component
}
\description{
generate k sample permutation
}
