% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{movingAverageGen}
\alias{movingAverageGen}
\title{generate data by moving average model}
\usage{
movingAverageGen(n, p, maParam, mu, innov = "normal")
}
\arguments{
\item{n}{n}

\item{p}{p}

\item{mu}{mean value}

\item{k}{MA order}

\item{rho}{MA parameter}
}
\description{
generate data by moving average model
}
