% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{randomizationTest}
\alias{randomizationTest}
\title{randomization test}
\usage{
randomizationTest(GTStat, genGT, initGT, B = 100, alpha = 0.05,
  needPvalue = FALSE)
}
\arguments{
\item{GTStat}{a function of random group trasformation. It calculate the randomization test statistic}

\item{genGT}{a function to generate random group transformation}

\item{initGT}{the identity of group transformation}

\item{B}{randomization times}

\item{alpha}{test level}
}
\description{
randomization test
}
