% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{simLevelA14}
\alias{simLevelA14}
\title{simulation of the level of A_1,...,A_4}
\usage{
simLevelA14(n, p, M, B, innov = list(type = "MA", innov = "normal"),
  maParam = NULL, factorParam = NULL, needPvalue = FALSE)
}
\arguments{
\item{n}{n}

\item{p}{p}

\item{M}{simulate M times test procedure}

\item{B}{resample B times to get randomization test}

\item{k}{Moving average parameter}

\item{rho}{Moving average parameter}

\item{mu}{mean value}
}
\description{
simulation of the level of A_1,...,A_4
}
