% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{simRanCQ}
\alias{simRanCQ}
\title{simulation of randomization test and CQ test}
\usage{
simRanCQ(n, p, M, mu, sparse, B, SNR, innov = list(type = "MA", innov =
  "normal"), maParam = NULL, factorParam = NULL, needPvalue = FALSE)
}
\arguments{
\item{n}{n}

\item{p}{p}

\item{M}{simulate M times test procedure to compute empirical power}

\item{mu}{mean value}

\item{B}{resample B times to get randomization test}

\item{SNR}{the signal noise ratio we need, mean value will scale to fit this SNR}

\item{k}{Moving average parameter}

\item{rho}{Moving average parameter}
}
\description{
simulation of randomization test and CQ test
}
