% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{simTwoSample}
\alias{simTwoSample}
\title{simulation of two sample permutation test for spiked model}
\usage{
simTwoSample(n1, n2, p, M, sparse, B, SNR, modelParam, needPvalue = FALSE)
}
\arguments{
\item{p}{p}

\item{M}{simulate M times test procedure to compute empirical power}

\item{B}{resample B times to get randomization test}

\item{SNR}{the signal noise ratio we need, mean value will scale to fit this SNR}

\item{modelParam}{the parameters of spiked model, including theSigmaSq, V, D, r}

\item{n}{n}
}
\description{
simulation of two sample permutation test for spiked model
}
