% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{spikedTempStatistic}
\alias{spikedTempStatistic}
\title{generate the randomization statistic based on the new method for spiked model}
\usage{
spikedTempStatistic(n1, n2, r, theData)
}
\arguments{
\item{n1}{n1}

\item{n2}{n2}

\item{theData}{the data}
}
\value{
ranStat a function of random group transformation
}
\description{
generate the randomization statistic based on the new method for spiked model
}
