% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatistics.R
\name{twoSampleTempStatistic}
\alias{twoSampleTempStatistic}
\title{generate the randomization statistic based on Chen Qin's statistic}
\usage{
twoSampleTempStatistic(n1, n2, theData)
}
\arguments{
\item{n1}{n1}

\item{n2}{n2}

\item{theData}{the data}
}
\value{
ranStat a function of random group transformation
}
\description{
generate the randomization statistic based on Chen Qin's statistic
}
