% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{varFanFactorModel}
\alias{varFanFactorModel}
\title{calculate the population variance of Fan's factor model}
\usage{
varFanFactorModel(p, theParam)
}
\description{
calculate the population variance of Fan's factor model
}
