% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{varMovingAverage}
\alias{varMovingAverage}
\title{calculate the population variance of moving average model}
\usage{
varMovingAverage(p, k, rho)
}
\description{
calculate the population variance of moving average model
}
